perf: fga-eps-mds/2020.2-Lend.it#121 Modularize instance creation functions
