+++ manage.py
@@ -9,6 +9,7 @@ COV = coverage.coverage(
         "project/api/models.py",
         "project/api/__init__.py",
         "project/__init__.py",
+        "project/api/utils.py",
     ],
 )
 COV.start()
